feat: update copy and add pending withdraw UI
