+++ src/modules/project/pages/projectFunding/views/refundPayoutRsk/PayoutRsk.tsx
@@ -18,7 +18,7 @@ import { Modal } from '@/shared/components/layouts/Modal.tsx'
 import { SkeletonLayout } from '@/shared/components/layouts/SkeletonLayout.tsx'
 import { Body } from '@/shared/components/typography/Body.tsx'
 import { Feedback, FeedBackVariant } from '@/shared/molecules/Feedback.tsx'
-import { getMempoolSpaceUrl } from '@/shared/utils/external/mempool.ts'
+import { getMempoolSpaceUrl, getRootstockBlockscoutUrl } from '@/shared/utils/external/mempool.ts'
 import { useCopyToClipboard } from '@/shared/utils/hooks/useCopyButton.ts'
 import {
   PaymentStatus,
@@ -363,12 +363,12 @@ export const PayoutRsk: React.FC<PayoutRskProps> = ({
   const waitingNotice = isWaitingClaimReady ? (
     t('Your funds are ready to be claimed')
   ) : (
-    <Trans i18nKey="We are waiting for the Bitcoin lockup transaction to be confirmed before you can claim the funds. You can check its status on <1>mempool.space</1>.">
+    <Trans i18nKey="We are waiting for the Bitcoin lockup transaction to be confirmed before you can claim the funds. You can check its status on <1>rootstock.blockscout.com</1>.">
       {
         'We are waiting for the Bitcoin lockup transaction to be confirmed before you can claim the funds. You can check its status on '
       }
-      <ChakraLink href={getMempoolSpaceUrl(lockTxId || '')} textDecoration="underline" isExternal>
-        {'mempool.space'}
+      <ChakraLink href={getRootstockBlockscoutUrl(lockTxId || '')} textDecoration="underline" isExternal>
+        {'rootstock.blockscout.com'}
       </ChakraLink>
       {'.'}
     </Trans>
